Consolidated the RackspaceIdentityBasicAuthHandler/Factory into the RackspaceIdentityBasicAuthFilter. Consolidated the tests in a similar fashion.
